Changed UsageSimulationBehavior contracts + added Postconditions class.
Because the interface that was removed in the last commit did not document well the contract, the events are now directly specified.

Furthermore, a new helper class has been implemented to check postconditions, especially when it comes to the events that should be published. These are typically asserted and hence used for verification.